Move weapon's component to separate package
