+++ source/core/src/main/com/csse3200/game/components/weapon/AllFiringLogic.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.components.player;
+package com.csse3200.game.components.weapon;
 
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.math.GridPoint2;
